+++ app/controllers/koi/application_controller.rb
@@ -1,5 +1,3 @@
-require 'garb'
-
 module Koi
   class ApplicationController < ActionController::Base
     helper :all
@@ -15,22 +13,9 @@ module Koi
     end
 
     def clear_cache
-      Rails.logger.warn("[CACHE CLEAR] - Clearning entire cache manually by #{current_admin} request")
+      Rails.logger.warn("[CACHE CLEAR] - Cleaning entire cache manually by #{current_admin} request")
       Rails.cache.clear
       redirect_to :back
-    end
-
-    def index
-      models = has_crud_models
-        .select { |h| h.crud.settings[:admin][:reportable] }
-        .select { |h| h.crud.settings[:admin][:overviews] }
-
-      @report_data = []
-
-      models.each do |model|
-        overviews = model.crud.settings[:admin][:overviews]
-        @report_data = @report_data | Reports::Reporting.generate_report(overviews, [], model)
-      end
     end
 
     protected