+++ app/models/asset.rb
@@ -1,4 +1,5 @@
 class Asset < ActiveRecord::Base
+  include Koi::Model
 
   has_crud paginate: false, settings: false