+++ app/models/concerns/koi/model.rb
@@ -1,0 +1,12 @@
+module Koi
+  module Model
+    extend ActiveSupport::Concern
+
+    included do
+      include Exportable
+      include HasSettings
+      include HasNavigation
+      include HasCrud::ActiveRecord
+    end
+  end
+end