+++ app/models/url_rewrite.rb
@@ -1,4 +1,5 @@
 class UrlRewrite < ActiveRecord::Base
+  include Koi::Model
 
   has_crud