+++ History.md
@@ -3,6 +3,7 @@
 
  * Removed generators and templates
  * Removed reports
+ * Removed Garb
 
 2.4 / 2020-08-04
 ==================