+++ History.md
@@ -2,6 +2,8 @@
 ==================
 
  * Removed generators and templates
+ * Removed reports
+ * Removed Garb
 
 2.4 / 2020-08-04
 ==================