+++ History.md
@@ -13,6 +13,7 @@
  * Removed rails-observers dependency
  * Removed active_model_serializers dependency
  * Removed mime-types dependency
+ * Removed redis dependency
 
 2.4 / 2020-08-04
 ==================