+++ History.md
@@ -16,6 +16,7 @@
  * Removed redis dependency
  * Removed uglifier dependency
  * Removed countries dependency
+ * Removed responders dependency
 
 2.4 / 2020-08-04
 ==================