+++ History.md
@@ -17,6 +17,7 @@
  * Removed uglifier dependency
  * Removed countries dependency
  * Removed responders dependency
+ * Removed sendgrid dependency
 
 2.4 / 2020-08-04
 ==================